+++ NEWS.md
@@ -1,8 +1,10 @@
 # phenoptr 0.1.6.9000
 
+Bug fixes:
 - `read_cell_seg_data` recognizes and correctly reads inForm data 
   which uses comma as the decimal separator (#8).
-  
+- Fix SpatialDistributionReport to work with more than 8 phenotypes (#6).
+
 # phenoptr 0.1.6.1
 **2019-04-22**